fix: cache session lookup so navigation doesn't block on the network
- Root beforeLoad re-runs on every navigation (and hover preload); it
  awaited an uncached getSession() round-trip each time, which was the
  visible navbar click lag
- Session now read via queryClient.ensureQueryData with 5-min staleTime
  (new src/lib/session.ts); repeat navigations resolve from cache
- Cache is dropped where the session changes: logout (user-menu),
  onboarding profile save (use-save-profile), Polar purchase success page
- AppHeader subscribes to pathname only instead of full router state,
  cutting per-transition header re-renders

Refs #231
