+++ src/features/onboarding/hooks/use-save-profile.ts
@@ -1,10 +1,13 @@
-import { useMutation } from "@tanstack/react-query"
+import { useMutation, useQueryClient } from "@tanstack/react-query"
 
 import { authClient } from "@/lib/auth-client"
+import { SESSION_QUERY_KEY } from "@/lib/session"
 import type { ProfileUpdateInput } from "@/server/users/users.type"
 
-export const useSaveProfile = () =>
-	useMutation({
+export const useSaveProfile = () => {
+	const queryClient = useQueryClient()
+
+	return useMutation({
 		mutationFn: async ({
 			username,
 			leetcodeHandle,
@@ -17,4 +20,10 @@ export const useSaveProfile = () =>
 			})
 			if (error) throw new Error(error.message ?? "Please try again.")
 		},
+		// The _authenticated guard reads username/leetcodeHandle from the cached
+		// session — drop it so the post-onboarding navigation sees the new values.
+		onSuccess: () => {
+			queryClient.removeQueries({ queryKey: SESSION_QUERY_KEY })
+		},
 	})
+}
