+++ src/routes/__root.tsx
@@ -3,7 +3,6 @@ import { type QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { ReactQueryDevtoolsPanel } from "@tanstack/react-query-devtools"
 import { createRootRouteWithContext, HeadContent, Scripts } from "@tanstack/react-router"
 import { TanStackRouterDevtoolsPanel } from "@tanstack/react-router-devtools"
-import { createIsomorphicFn } from "@tanstack/react-start"
 import { Analytics } from "@vercel/analytics/react"
 import { SpeedInsights } from "@vercel/speed-insights/react"
 import type { ReactNode } from "react"
@@ -12,24 +11,16 @@ import { Toaster } from "sileo"
 import { ErrorPage } from "@/components/error"
 import { NotFoundPage } from "@/components/not-found"
 import { Spinner } from "@/components/ui/spinner"
-import { authClient } from "@/lib/auth-client"
 import { getQueryClient } from "@/lib/query-client"
+import { sessionQueryOptions } from "@/lib/session"
 import appCss from "../styles.css?url"
 
-const fetchSession = createIsomorphicFn()
-	.client(async () => {
-		const { data } = await authClient.getSession()
-		return data
-	})
-	.server(async () => {
-		const { auth } = await import("@/server/lib/auth")
-		const { getRequestHeaders } = await import("@tanstack/react-start/server")
-		return await auth.api.getSession({ headers: getRequestHeaders() })
-	})
-
 export const Route = createRootRouteWithContext<{ queryClient: QueryClient }>()({
-	beforeLoad: async () => {
-		const session = await fetchSession()
+	// beforeLoad re-runs on every navigation (and every hover preload). Reading
+	// the session through the query cache keeps repeat navigations off the
+	// network — the uncached fetch here was the navbar lag in #231.
+	beforeLoad: async ({ context }) => {
+		const session = await context.queryClient.ensureQueryData(sessionQueryOptions)
 		return { session }
 	},
 	head: () => ({
